refactor: improve dependency management.
So far, each project in keybroker-demo was specifying its own package
dependencies, including the version pinning and features. This is
undesirable as over time, it means that projects' dependencies might
unnecessarily diverge. Cargo's dependency resolver is doing a good
job... but hiding the mess under the carpet does not seem like a good
engineering approach when there are available solutions.

Instead, let's have all the package dependencies in the top
Cargo.toml, and let the individual projects refer to them. This gives us
a central place where to manage the dependencies from. This may become
important in the future, e.g. if we have to update some package for
security reasons for example.

Signed-off-by: Arnaud de Grandmaison <arnaud.degrandmaison@arm.com>